final adjustments before error check
After a lot of tinkering and adjusting I got most of what I think is alright, for some reason it stopped showing background color on firefox and it would be able to go past what was set which is not supposed to happen. Another thing is that it works well on google chrome browsers. I tried to get the checkbox to move to the side of "I am over 18 years old" but didn't have alot of luck. Besides I have another lab I have to finish as well as study for other midterms as well as write my case study.